// src/enums/users-table/user-db.enum.ts
export enum UserDbEnum {
  Id = 'id',
  Role = 'role',
  Nickname = 'nickname',
  Email = 'email',
  LastsName = 'lastsName',
  FirsName = 'firsName',
  ServiceIdsStr = 'serviceIdsStr',
  CityIdsStr = 'cityIdsStr',
  Phone = 'phone',
  Gender = 'gender',
  Birthday = 'birthday',
  Avatar = 'avatar',
  InfoYourself = 'infoYourself',
}